Move basic pushdown types to PredicatePushdownController
